refactor: corrige estrutura
