+++ model/interfaceData.ts
@@ -1,0 +1,8 @@
+export interface NewRow {
+    nome: string,
+    peso: number,
+    valor: number,
+    quant: number
+}
+
+
